bitna/web-gui/requirements.txt
nicegui
fastapi
uvicorn[standard]
pymysql
python-dotenv